+++ android/src/main/java/dev/barooni/capacitor/calendar/models/data/EventDeleteInfo.kt
@@ -3,5 +3,6 @@ package dev.barooni.capacitor.calendar.models.data
 data class EventDeleteInfo(
     val dtStart: Long?,
     val originalId: Long?,
+    val originalInstanceTime: Long?,
     val rrule: String?,
 )